+++ regression/reference/error-ast/wrapferror.f
@@ -13,6 +13,7 @@
 ! splicer begin file_top
 ! splicer end file_top
 module error_mod
+    use iso_c_binding, only : C_INT, C_NULL_PTR, C_PTR
     ! splicer begin module_use
     ! splicer end module_use
     implicit none
@@ -20,12 +21,86 @@ module error_mod
     ! splicer begin module_top
     ! splicer end module_top
 
+    ! helper capsule_data_helper
+    type, bind(C) :: ERR_SHROUD_capsule_data
+        type(C_PTR) :: addr = C_NULL_PTR  ! address of C++ memory
+        integer(C_INT) :: idtor = 0       ! index of destructor
+    end type ERR_SHROUD_capsule_data
+
+    type class1
+        type(ERR_SHROUD_capsule_data) :: cxxmem
+        ! splicer begin class.Class1.component_part
+        ! splicer end class.Class1.component_part
+    contains
+        procedure :: get_instance => class1_get_instance
+        procedure :: set_instance => class1_set_instance
+        procedure :: associated => class1_associated
+        ! splicer begin class.Class1.type_bound_procedure_part
+        ! splicer end class.Class1.type_bound_procedure_part
+    end type class1
+
+    interface operator (.eq.)
+        module procedure class1_eq
+    end interface
+
+    interface operator (.ne.)
+        module procedure class1_ne
+    end interface
+
     ! splicer begin additional_declarations
     ! splicer end additional_declarations
 
 contains
 
+    ! Return pointer to C++ memory.
+    function class1_get_instance(obj) result (cxxptr)
+        use iso_c_binding, only: C_PTR
+        class(class1), intent(IN) :: obj
+        type(C_PTR) :: cxxptr
+        cxxptr = obj%cxxmem%addr
+    end function class1_get_instance
+
+    subroutine class1_set_instance(obj, cxxmem)
+        use iso_c_binding, only: C_PTR
+        class(class1), intent(INOUT) :: obj
+        type(C_PTR), intent(IN) :: cxxmem
+        obj%cxxmem%addr = cxxmem
+        obj%cxxmem%idtor = 0
+    end subroutine class1_set_instance
+
+    function class1_associated(obj) result (rv)
+        use iso_c_binding, only: c_associated
+        class(class1), intent(IN) :: obj
+        logical rv
+        rv = c_associated(obj%cxxmem%addr)
+    end function class1_associated
+
+    ! splicer begin class.Class1.additional_functions
+    ! splicer end class.Class1.additional_functions
+
     ! splicer begin additional_functions
     ! splicer end additional_functions
+
+    function class1_eq(a,b) result (rv)
+        use iso_c_binding, only: c_associated
+        type(class1), intent(IN) ::a,b
+        logical :: rv
+        if (c_associated(a%cxxmem%addr, b%cxxmem%addr)) then
+            rv = .true.
+        else
+            rv = .false.
+        endif
+    end function class1_eq
+
+    function class1_ne(a,b) result (rv)
+        use iso_c_binding, only: c_associated
+        type(class1), intent(IN) ::a,b
+        logical :: rv
+        if (.not. c_associated(a%cxxmem%addr, b%cxxmem%addr)) then
+            rv = .true.
+        else
+            rv = .false.
+        endif
+    end function class1_ne
 
 end module error_mod