+++ regression/run/cxxlibrary/cxxlibrary.hpp
@@ -66,5 +66,27 @@ struct nested {
     nested **child;
 };
 
+//----------------------------------------------------------------------
+
+class Class1
+{
+public:
+    int m_length;
+    Class1() : m_length(99) {};
+
+    // Test fortran_generic with default arguments.
+    int check_length(int length = 1)
+    {
+        return length;
+    };
+
+    // test return_this
+    Class1* declare(int flag, int length = 1)
+    {
+        m_length = length;
+        return this;
+    };
+};
+
 #endif // CXXLIBRARY_H