+++ shroud/generate.py
@@ -158,7 +158,9 @@ class VerifyAttrs(object):
         if ast.typemap is None:
             print("XXXXXX typemap is None")
         if ast.typemap.sgroup == "shadow":
-            if options.C_shadow_result:
+            if node.return_this:
+                meta["api"] = "this"
+            elif options.C_shadow_result:
                 meta["api"] = "capptr"
             else:
                 meta["api"] = "capsule"
@@ -671,6 +673,7 @@ class GenFunctions(object):
           process_class
             add_var_getter_setter
             define_function_suffix
+              append_function_index
               has_default_args
               template_function
               define_fortran_generic_functions
@@ -895,6 +898,7 @@ class GenFunctions(object):
         fcn.wrap.lua = False
         fcn.wrap.python = False
         fcn._generated = "getter/setter"
+        fcn._generated_path.append("getter/setter")
 
         ##########
         # setter
@@ -928,6 +932,7 @@ class GenFunctions(object):
         fcn.wrap.lua = False
         fcn.wrap.python = False
         fcn._generated = "getter/setter"
+        fcn._generated_path.append("getter/setter")
 
     def instantiate_all_classes(self, node):
         """Instantate all class template_arguments recursively.
@@ -967,6 +972,8 @@ class GenFunctions(object):
                 orig_typemap = cls.typemap
                 if orig_typemap.cxx_instantiation is None:
                     orig_typemap.cxx_instantiation = {}
+                for function in cls.functions:
+                    self.append_function_index(function)
                 # Replace class with new class for each template instantiation.
                 # targs -> ast.TemplateArgument
                 for i, targs in enumerate(cls.template_arguments):
@@ -1079,6 +1086,7 @@ class GenFunctions(object):
         node = cls.add_function(name, ast, options=opt)
         node.declgen = node.ast.gen_decl()
         node._generated = "struct_as_class_ctor"
+        node._generated_path.append("struct_as_class_ctor")
 
     def process_class(self, parent, cls):
         """Process variables and functions for a class/struct.
@@ -1203,14 +1211,7 @@ class GenFunctions(object):
                     if not function.fmtdict.inlocal("function_suffix"):
                         function.fmtdict.function_suffix = "_{}".format(i)
 
-        # return_this
-        ordered2 = []
-        for method in ordered_functions:
-            ordered2.append(method)
-            if method.return_this:
-                self.process_return_this(method, ordered2)
-                
-        ordered3 = self.define_fortran_generic_functions(ordered2)
+        ordered3 = self.define_fortran_generic_functions(ordered_functions)
         ordered4 = self.define_bufferify_functions(ordered3)
 
         self.gen_functions_decl(ordered4)
@@ -1246,6 +1247,7 @@ class GenFunctions(object):
             self.append_function_index(new)
 
             new._generated = "cxx_template"
+            new._generated_path.append("cxx_template")
 
             fmt = new.fmtdict
             if targs.fmtdict:
@@ -1327,6 +1329,7 @@ class GenFunctions(object):
         self.append_function_index(new)
 
         new._generated = "cxx_template"
+        new._generated_path.append("cxx_template")
 
         new.cxx_template = {}
         #        fmt.CXX_template = targs.instantiation   # ex. <int>
@@ -1448,6 +1451,7 @@ class GenFunctions(object):
             ordered_functions.append(new)
             self.append_function_index(new)
             new._generated = "fortran_generic"
+            new._generated_path.append("fortran_generic")
             fmt = new.fmtdict
             # XXX append to existing suffix
             if generic.fmtdict:
@@ -1455,12 +1459,17 @@ class GenFunctions(object):
             fmt.function_suffix = fmt.function_suffix + generic.function_suffix
             new.fortran_generic = {}
             new.wrap.assign(fortran=True)
+            if len(new.ast.declarator.params) != len(generic.decls):
+                raise RuntimeError("internal: generic_function: length mismatch: "
+                                   + node.name)
             new.ast.declarator.params = generic.decls
 
             # Try to call original C function if possible.
             # All arguments are native scalar.
             need_wrapper = False
-            if new.ast.declarator.is_indirect():
+            if new.return_this:
+                pass
+            elif new.ast.declarator.is_indirect():
                 need_wrapper = True
             
             for arg in new.ast.declarator.params:
@@ -1500,6 +1509,10 @@ class GenFunctions(object):
           void func(int i, int j)
         In Fortran, these are added to a generic interface.
 
+        It is also necessary to trim fortran_generic.
+        For example, func() will not have any generic variations
+        since it has no arguments.
+
         Args:
             node -
             ordered_functions -
@@ -1522,8 +1535,11 @@ class GenFunctions(object):
             new = node.clone()
             self.append_function_index(new)
             new._generated = "has_default_arg"
+            new._generated_path.append("has_default_arg")
             del new.ast.declarator.params[i:]  # remove trailing arguments
             new._has_default_arg = False
+            if node.fortran_generic:
+                new.fortran_generic = ast.trim_fortran_generic_decls(node.fortran_generic, i)
             # Python and Lua both deal with default args in their own way
             new.wrap.assign(c=True, fortran=True)
             fmt = new.fmtdict
@@ -1573,43 +1589,6 @@ class GenFunctions(object):
                 attrs[name] = c_attrs[name]
                 del c_attrs[name]
 
-    def process_return_this(self, node, ordered_functions):
-        """Deal with return_this feature.
-
-        If a function is marked return_this, convert it into a 
-        subroutine for the C and Fortran wrappers.
-        Return this allows chaining of function calls.
-        For example in C++:   obj->doA()->doB();
-        Python:   obj.doA().doB()
-        However, there is no way to chain in C or Fortran.
-
-        Clone the function and wrap for C and Fortran.
-        Turn off C and Fortran wrapper on original node.
-        Remove the function result.
-        
-        Parameters
-        ----------
-        node : FunctionNode
-        ordered_functions : list of FunctionNode
-        """
-        if node.wrap.c == False and node.wrap.fortran == False:
-            return
-        new = node.clone()
-        ordered_functions.append(new)
-        self.append_function_index(new)
-        new._generated = "return_this"
-
-        # Only wrap for C and Fortran, transfer values from node.
-        new.wrap.clear()
-        new.wrap.c = node.wrap.c
-        new.wrap.fortran = node.wrap.fortran
-        node.wrap.c = False
-        node.wrap.fortran = False
-
-        # Do not return C++ this instance.
-        new.ast.set_return_to_void()
-        new.ast.declarator.metaattrs["intent"] = "subroutine"
-    
     def arg_to_CFI(self, node, ordered_functions):
         """Look for functions which can use TS29113
         Futher interoperability with C.
@@ -1716,6 +1695,7 @@ class GenFunctions(object):
 
         generated_suffix = "cfi"
         C_new._generated = "arg_to_cfi"
+        C_new._generated_path.append("arg_to_cfi")
         C_new.splicer_group = "cfi"
         if need_buf_result:
             C_new.ast.declarator.metaattrs["api"] = need_buf_result
@@ -1901,6 +1881,7 @@ class GenFunctions(object):
 
         generated_suffix = "buf"
         C_new._generated = "arg_to_buffer"
+        C_new._generated_path.append("arg_to_buffer")
         C_new.splicer_group = "buf"
         if need_buf_result:
             C_new.ast.declarator.metaattrs["api"] = need_buf_result