+++ CHANGELOG.md
@@ -45,6 +45,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed attribute *+cdesc*. Replace by *+api(cdesc)*
 - Removed *f_module_line* and *c_module_line* from both statements and Typemap.
 - Remove *f_imports* from statements.
+- No longer create a new function for return_this.
+  Instead set api(this) and let statements do the right thing.
 
 ## v0.13.0 - 2023-08-23
 ### Added